backend/backend/app/backend/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
# Agrega cualquier otra dependencia que uses (ej. Flask-Bcrypt para hashing de contraseñas)